Moved labwhere client helper into support directory
