+++ spec/support/lab_where_client_helper.rb
